fixing price amount change uint to float32

+++ infrastructure/now_payments_response/get_payment_status_response.go
@@ -11,7 +11,7 @@ type GetPaymentStatusResponse struct {
 	PriceAmount      float32   `json:"price_amount"`
 	PriceCurrency    string    `json:"price_currency"`
 	PayAmount        float32   `json:"pay_amount"`
-	ActualPaid       uint      `json:"actually_paid"`
+	ActualPaid       float32   `json:"actually_paid"`
 	PayCurrency      string    `json:"pay_currency"`
 	OrderId          *string   `json:"order_id"`
 	OrderDescription *string   `json:"order_description"`
